requirements.txt
Django==5.1.1
django-environ==0.11.2
docker==7.1.0
ovh==1.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
urllib3==2.2.3

pip~=24.2
pillow~=11.0.0
prometheus-api-client~=0.5.5